Extract token utilities
